--- a/benchmarks/shard.yml
+++ b/benchmarks/shard.yml
@@ -8,9 +8,6 @@ dependencies:
   artanis:
     github: ysbaddaden/artanis
     branch: master
-  beryl:
-    github: luislavena/crystal-beryl
-    branch: master
   mime:
     github: spalger/crystal-mime
     branch: master