--- a/src/beryl.cr
+++ b/src/beryl.cr
@@ -1,4 +1,4 @@
-require "./beryl/*"
+require "./beryl/**"
 
 # Action-focused HTTP routing library for Crystal.
 module Beryl