Implementation of a Radix Tree as alternative routing mechanism.

This implementation will replace current linear route matcher
(Array+Regexp) which has shown slow performance.

- Generic `Node` that supports any payload to be associated with it
- `Tree` implementation optimized for path matching using Char::Reader
- `Result` object which provides accumulative results of finding
  a path on a `Tree`

This code has been inspired and adapted from julienschmidt/httprouter [1],
but reduced and optimized to Crystal capabilities.

[1] https://github.com/julienschmidt/httprouter

---

Use Radix Tree for Routing

- Remove `Route` and `CompiledPattern`
- Use multiple trees for each request method
- Direct mapping to action within Router

On synthetic benchmarks (compare-frameworks) things have improved:

From:

```
amethyst (req_middle)   38.9k (± 0.61%)  2.70× slower
 artanis (req_middle)  96.02k (± 0.65%)  1.09× slower
   beryl (req_middle)   7.64k (± 1.04%) 13.75× slower
 routing (req_middle) 105.13k (± 0.77%)       fastest

amethyst (req_bottom)  20.33k (± 0.68%)  4.78× slower
 artanis (req_bottom)   97.1k (± 0.94%)       fastest
   beryl (req_bottom)    4.9k (± 0.88%) 19.83× slower
 routing (req_bottom)   7.36k (± 0.73%) 13.19× slower
```

To:

```
amethyst (req_middle)  39.29k (± 0.51%)  4.39× slower
 artanis (req_middle)  99.14k (± 0.51%)  1.74× slower
   beryl (req_middle) 172.65k (± 0.76%)       fastest
 routing (req_middle)  109.3k (± 0.69%)  1.58× slower

amethyst (req_bottom)   21.1k (± 1.53%) 10.56× slower
 artanis (req_bottom)  93.15k (± 1.72%)  2.39× slower
   beryl (req_bottom) 222.93k (± 0.69%)       fastest
 routing (req_bottom)   7.67k (± 0.47%) 29.06× slower
```

---

Both benchmarks shows increased performance in middle request
(on big trees) and bottom (last defined route)

However, there is no major improvement in regular requests:

From:

```
amethyst (req_root)   71.3k (± 0.46%)  13.59× slower
 artanis (req_root)  969.2k (± 0.94%)        fastest
   beryl (req_root)    261k (± 0.95%)   3.71× slower
 routing (req_root)   8.39k (± 0.64%) 115.51× slower

amethyst (req_id)  56.96k (± 0.68%)  9.43× slower
 artanis (req_id) 537.33k (± 0.92%)       fastest
   beryl (req_id) 125.74k (± 1.07%)  4.27× slower
 routing (req_id) 256.28k (± 0.83%)  2.10× slower
```

To:

```
amethyst (req_root)   70.7k (± 0.71%)  13.96× slower
 artanis (req_root) 987.37k (± 0.67%)        fastest
   beryl (req_root) 279.59k (± 1.02%)   3.53× slower
 routing (req_root)   8.82k (± 0.61%) 111.91× slower

amethyst (req_id)  56.94k (± 0.74%)  9.95× slower
 artanis (req_id) 566.57k (± 1.19%)       fastest
   beryl (req_id) 186.81k (± 1.20%)  3.03× slower
 routing (req_id) 264.53k (± 0.45%)  2.14× slower
```

---

Which shows Artanis still outperforms all the compared frameworks.