@luislavena 2015-11-13
I've added one last benchmark script to compare performance of master vs this branch.

Running `bench-github` (compiled in release mode), produces the following when tested against `wrk`:

## master

```
$ ./wrk -t 2 http://localhost:9292/user/repos
Running 10s test @ http://localhost:9292/user/repos
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   268.92us  117.17us   5.39ms   87.01%
    Req/Sec    18.87k     1.04k   19.73k    82.67%
  379271 requests in 10.10s, 31.47MB read
Requests/sec:  37553.26
Transfer/sec:      3.12MB

$ ./wrk -t 2 http://localhost:9292/repos/luislavena/crystal-beryl/stargazers
Running 10s test @ http://localhost:9292/repos/luislavena/crystal-beryl/stargazers
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   216.94us   96.68us   5.66ms   91.97%
    Req/Sec    23.39k     0.98k   24.26k    85.15%
  470134 requests in 10.10s, 39.01MB read
Requests/sec:  46550.49
Transfer/sec:      3.86MB
```

## PR

```
$ ./wrk -t 2 http://localhost:9292/user/repos
Running 10s test @ http://localhost:9292/user/repos
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   144.56us   73.30us   3.54ms   89.26%
    Req/Sec    35.25k     3.43k   72.10k    90.55%
  704971 requests in 10.10s, 58.49MB read
Requests/sec:  69799.05
Transfer/sec:      5.79MB

$ ./wrk -t 2 http://localhost:9292/repos/luislavena/crystal-beryl/stargazers
Running 10s test @ http://localhost:9292/repos/luislavena/crystal-beryl/stargazers
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   183.90us   84.56us   4.65ms   86.75%
    Req/Sec    27.60k     2.49k   55.36k    90.55%
  551993 requests in 10.10s, 45.80MB read
Requests/sec:  54656.76
Transfer/sec:      4.53MB
```

There is a clear improvement in fixed routes (eg. `/user/repos`) and slight but not so great improvement on named segment routes (eg. `/repos/:owner/:repo/stargazers`).

I think this is good now to be merged.